pyteststudy/requirement.txt
pytest-html
pytest-xdist
pytest-rerunfailures




